sorry  i was very stuck, i needed to utilize at
i didnt copy everthing but it did provide a few pieces of code that i did indeed kinda rewrite to work for this. in particualar this fancy piece of code.

    const ad = document.querySelector('.adBanner');
    if (ad) ad.style.display = account.premium ? 'none' : 'flex';

and i think also this
  async function togglePremium() {
    await fetch('/togglePremium', { method: 'POST' });
    refreshPremiumUI();
  }

i will make it a point to understand all of the code and give refernces to my research.